refactor: keep quantity serialization in k8s types
